+++ examples/foxes/single_state/run.py
@@ -47,13 +47,11 @@ if __name__ == "__main__":
         default=["B_K1", "CH_K2"],
         nargs="+",
     )
-    parser.add_argument(
-        "-g", "--ground", help="switch on ground mirror", action="store_true"
-    )
     parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
     parser.add_argument(
         "-p", "--pwakes", help="The partial wakes models", default=None, nargs="+"
     )
+    parser.add_argument("-g", "--grounds", help="The ground models", default=None, nargs="+")
     parser.add_argument("-f", "--frame", help="The wake frame", default="rotor_wd")
     parser.add_argument("-v", "--var", help="The plot variable", default=FV.WS)
     parser.add_argument(
@@ -105,12 +103,6 @@ if __name__ == "__main__":
             farm, args.layout, turbine_models=args.tmodels + [ttype.name]
         )
 
-    # optionally add wake ground mirror:
-    if args.ground:
-        mirrors = {w: [0] for w in args.wakes}
-    else:
-        mirrors = {}
-
     # create algorithm
     Algo = foxes.algorithms.Iterative if args.iterative else foxes.algorithms.Downwind
     algo = Algo(
@@ -120,7 +112,7 @@ if __name__ == "__main__":
         wake_models=args.wakes,
         wake_frame=args.frame,
         partial_wakes=args.pwakes,
-        wake_mirrors=mirrors,
+        ground_models=args.grounds,
         mbook=mbook,
         chunks=None,
         verbosity=1,