+++ foxes/models/ground_models/__init__.py
@@ -1,0 +1,2 @@
+from .no_ground import NoGround
+from .wake_mirror import WakeMirror, GroundMirror