+++ foxes/models/partial_wakes/rotor_points.py
@@ -9,7 +9,6 @@ class RotorPoints(PartialWakesModel):
     :group: models.partial_wakes
 
     """
-
     def get_wake_points(self, algo, mdata, fdata):
         """
         Get the wake calculation points, and their