+++ foxes/models/partial_wakes/segregated.py
@@ -20,7 +20,6 @@ class PartialSegregated(PartialWakesModel):
     :group: models.partial_wakes
 
     """
-
     def __init__(self, rotor_model):
         """
         Constructor.