+++ foxes/models/wake_models/induction/__init__.py
@@ -6,3 +6,4 @@ from .rankine_half_body import RankineHalfBody
 from .rathmann import Rathmann
 from .self_similar import SelfSimilar
 from .self_similar2020 import SelfSimilar2020
+from .vortex_sheet import VortexSheet